chore: split components into individual code files
