+++ tools/serverpod_cli/lib/src/commands/tui/components/shimmer.dart
