+++ tools/serverpod_cli/lib/src/commands/tui/components/spinner.dart
